Move table-cell-radio to generic folder
