+++ src/frontend/app/shared/components/list/list-table/table-cell-radio/table-cell-radio.component.html
