+++ src/frontend/app/shared/components/list/list-table/table-cell-radio/table-cell-radio.component.spec.ts
@@ -1,15 +1,15 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TableCellRadioComponent } from './table-cell-radio.component';
-import { CoreModule } from '../../../../../../core/core.module';
-import { SharedModule } from '../../../../../shared.module';
-import { ApplicationService } from '../../../../../../features/applications/application.service';
-import { ApplicationServiceMock } from '../../../../../../test-framework/application-service-helper';
+import { CoreModule } from '../../../../../core/core.module';
+import { SharedModule } from '../../../../shared.module';
+import { ApplicationService } from '../../../../../features/applications/application.service';
+import { ApplicationServiceMock } from '../../../../../test-framework/application-service-helper';
 import {
   EntityInfo,
   APIResource
-} from '../../../../../../store/types/api.types';
-import { IListDataSource } from '../../../data-sources-controllers/list-data-source-types';
+} from '../../../../../store/types/api.types';
+import { IListDataSource } from '../../data-sources-controllers/list-data-source-types';
 
 describe('TableCellRadioComponent', () => {
   let component: TableCellRadioComponent<any>;
